improve message config UI - auto increment id, copy data type
When adding a tag in the message configuration, the data type is now
copied from the previous tag and the tag id is incremented.

+++ src/main/java/com/chitek/ignition/drivers/generictcp/meta/config/TagConfig.java
@@ -36,7 +36,7 @@ public class TagConfig implements Serializable {
 	// Set a default datatype for new items
 	public TagConfig()
 	{
-		dataType = BinaryDataType.Byte;
+		dataType = BinaryDataType.Int16;
 		size = 1;
 	}
